Split rsloaderCreateExec into 2 functions.
In some scenerio, such as cross-compilation, we have to
load the ELF object file but not relocate them.  Thus,
we are going to provide 2 functions:

  1. rsloaderLoadExecutable
  2. rsloaderRelocateExecutable

So that we can load the executable WITHOUT relocation.

The behavior of rsloaderCreateExec is remained unchanged
in this commit, though rsloaderCreateExec will become
the composition of these 2 functions.

Change-Id: Idb64aa77ac53161aa95879ccad8f46ce56c1d993